test: expand coverage for mindmap, CLI, and graph
- Refactored Mind Map tree construction and ARIA logic into `src/lib/mindmap-tree.ts`.
- Added comprehensive unit and component tests for Mind Map.
- Implemented CLI tests for database initialization and commands using a temporary SQLite database.
- Refactored Graph layout and construction logic into `src/lib/graph-data.ts`.
- Added unit and component tests for the Knowledge Graph.
- Added tests for `useFocusTrap` hook, `MarkdownRenderer`, and `DbProvider`.
- Updated test setup to support perf markings in a test environment.
- Fixed `MarkdownRenderer` to correctly render code blocks by allowing `pre` and `code` tags in `sanitizeHtml`.
- Verified all tests pass locally.
