+++ src/features/mindmap/MindMapView.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useRef, useState, useMemo, useCallback } from 'react';
-import MindElixir, { type MindElixirInstance } from 'mind-elixir';
+import MindElixir, { type MindElixirData, type MindElixirInstance } from 'mind-elixir';
 import type { Entity, Link } from '../../lib/validation';
-import { buildTree, addAriaAttributesToContainer } from '../../lib/mindmap-tree';
 import { repository } from '../../db/repository';
 import { logger } from '../../lib/logger';
 import { upsertToSearchIndex } from '../../lib/search';
@@ -17,6 +16,50 @@ interface Props {
   entities: Entity[];
   links: Link[];
   onEntityClick?: (entityId: string) => void;
+}
+
+function buildTree(
+  currentId: string,
+  depth: number,
+  maxDepth: number,
+  entities: Entity[],
+  links: Link[],
+  relationFilter: string,
+): MindElixirData['nodeData'] | null {
+  const entity = entities.find(e => e.id === currentId);
+  if (!entity || depth > maxDepth) return null;
+
+  const childrenLinks = links.filter(l =>
+    l.source_id === currentId &&
+    (relationFilter === 'all' || l.relation === relationFilter)
+  );
+
+  return {
+    id: entity.id || `node-${Math.random()}`,
+    topic: entity.name,
+    children: childrenLinks
+      .map(l => buildTree(l.target_id, depth + 1, maxDepth, entities, links, relationFilter))
+      .filter((n): n is MindElixirData['nodeData'] => n !== null)
+  };
+}
+
+function addAriaToNodes(container: HTMLElement): void {
+  const topics = container.querySelectorAll('me-tpc');
+  topics.forEach(tpc => {
+    const parent = tpc.closest('me-parent');
+    if (parent && !parent.hasAttribute('role')) {
+      parent.setAttribute('role', 'treeitem');
+      parent.setAttribute('aria-label', tpc.textContent?.trim() || 'Mind map node');
+    }
+  });
+}
+
+function addAriaAttributesToContainer(container: HTMLElement): void {
+  container.setAttribute('role', 'tree');
+  addAriaToNodes(container);
+  const nodeObserver = new MutationObserver(() => { addAriaToNodes(container); });
+  nodeObserver.observe(container, { childList: true, subtree: true });
+  setTimeout(() => { nodeObserver.disconnect(); }, 2000);
 }
 
 const MindMapView: React.FC<Props> = ({
